Added dynamic items and subitems for wizard sidebar
