+++ src/routes/WizardRoute/WizardLayout.tsx
@@ -6,11 +6,14 @@ import React, {
   useMemo,
   useState,
 } from 'react';
+import { defineMessages } from 'react-intl';
 
 import Wallet from '~frame/RouteLayouts/UserNavigation/Wallet';
 import { Header } from '~frame/Extensions/layouts';
 
 import WizardSidebar from './WizardSidebar';
+
+const displayName = 'routes.WizardRoute.WizardLayout';
 
 interface ColonyCreationFlowContextValues {
   currentStep: number;
@@ -31,6 +34,76 @@ export const useColonyCreationFlowContext = () => {
   return context;
 };
 
+const MSG = defineMessages({
+  sidebarTitle: {
+    id: `${displayName}.sidebarTitle`,
+    defaultMessage: 'Create your new Colony',
+  },
+  account: {
+    id: `${displayName}.account`,
+    defaultMessage: 'Account',
+  },
+  profile: {
+    id: `${displayName}.profile`,
+    defaultMessage: 'Profile',
+  },
+  create: {
+    id: `${displayName}.create`,
+    defaultMessage: 'Create',
+  },
+  details: {
+    id: `${displayName}.details`,
+    defaultMessage: 'Details',
+  },
+  nativeToken: {
+    id: `${displayName}.nativeToken`,
+    defaultMessage: 'Native Token',
+  },
+  confirmation: {
+    id: `${displayName}.confirmation`,
+    defaultMessage: 'Confirmation',
+  },
+  complete: {
+    id: `${displayName}.complete`,
+    defaultMessage: 'Complete',
+  },
+});
+
+const wizardSteps = [
+  {
+    itemStep: 0,
+    itemText: MSG.account,
+    subItems: [
+      {
+        itemStep: 0,
+        itemText: MSG.profile,
+      },
+    ],
+  },
+  {
+    itemStep: 1,
+    itemText: MSG.create,
+    subItems: [
+      {
+        itemStep: 1,
+        itemText: MSG.details,
+      },
+      {
+        itemStep: 2,
+        itemText: MSG.nativeToken,
+      },
+      {
+        itemStep: 3,
+        itemText: MSG.confirmation,
+      },
+    ],
+  },
+  {
+    itemStep: 4,
+    itemText: MSG.complete,
+  },
+];
+
 const WizardLayout: FC<PropsWithChildren> = ({ children }) => {
   const [currentStep, setCurrentStep] = useState(0);
   const colonyCreationFlowContextValues = useMemo(
@@ -42,7 +115,10 @@ const WizardLayout: FC<PropsWithChildren> = ({ children }) => {
     <ColonyCreationFlowContext.Provider value={colonyCreationFlowContextValues}>
       <div className="grid grid-cols-[280px,auto] gap-4 p-4">
         <aside className="sticky top-4 h-[calc(100vh-2rem)]">
-          <WizardSidebar />
+          <WizardSidebar
+            wizardSteps={wizardSteps}
+            sidebarTitle={MSG.sidebarTitle}
+          />
         </aside>
         <div className="">
           <Header />
