+++ src/routes/WizardRoute/WizardSidebar/index.ts
@@ -1,0 +1,1 @@
+export { default } from './WizardSidebar';
